src/requirements.txt
astroid==2.3.3
certifi==2019.11.28
chardet==3.0.4
Click==7.0
Flask==1.1.1
idna==2.8
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.10.3
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
pkg-resources==0.0.0
PyGithub==1.45.0
pylint==2.4.4
python-dotenv==0.10.5
pytz==2019.3
requests==2.22.0
six==1.14.0
typed-ast==1.4.1
urllib3==1.25.8
Werkzeug==0.16.0
wrapt==1.11.2
